% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{alpha_from_mi_om3}
\alias{alpha_from_mi_om3}
\title{Budyko relationship}
\usage{
alpha_from_mi_om3(mi)
}
\arguments{
\item{mi}{Numeric value with moisture index.}
}
\value{
Numeric value with alpha.
}
\description{
Budyko relationship, obtain alpha from moisture index (MI).
}
