% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{chi}
\alias{chi}
\title{Ratio of leaf-internal to ambient CO2 partial pressures (–)}
\usage{
chi(Tc, MI, co2, scale_factor = 101.325 * 10^-3)
}
\arguments{
\item{Tc}{Numeric value of temperature (°C).}

\item{MI}{Numeric value of moisture index (-).}

\item{co2}{Numeric value of CO2 partial pressure (umol/mol).}

\item{scale_factor}{Scale factor to transform the output, default =
101.325 Pa/ppm at standard sea level pressure.}
}
\value{
Numeric value of ratio of leaf-internal to ambient CO2 partial
pressures.
}
\description{
Ratio of leaf-internal to ambient CO2 partial pressures (–)
}
\keyword{internal}
