% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{convert_units.m2d}
\alias{convert_units.m2d}
\title{Convert units from monthly to daily}
\usage{
convert_units.m2d(
  filename,
  varid,
  timeid = "time",
  latid = "lat",
  lonid = "lon",
  FUN = `*`
)
}
\arguments{
\item{filename}{Filename for the netCDF input (relative or absolute path).}

\item{varid}{String with the main variable identifier.}

\item{timeid}{String with the time dimension identifier.}

\item{latid}{String with the latitude dimension identifier.}

\item{lonid}{String with the longitude dimension identifier.}

\item{FUN}{Infix function to perform the conversion.}
}
\description{
Convert units from monthly to daily
}
