% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{convert_units}
\alias{convert_units}
\title{Convert units from a netCDF file}
\usage{
convert_units(
  filename,
  varid,
  new_units,
  conv_factor,
  timeid = "time",
  latid = "lat",
  lonid = "lon",
  overwrite = TRUE,
  FUN = `*`
)
}
\arguments{
\item{filename}{Filename for the netCDF input (relative or absolute path).}

\item{varid}{String with the main variable identifier.}

\item{new_units}{String with the new units.}

\item{conv_factor}{Numeric vector or single value with the conversion factor.}

\item{timeid}{String with the time dimension identifier.}

\item{latid}{String with the latitude dimension identifier.}

\item{lonid}{String with the longitude dimension identifier.}

\item{overwrite}{Boolean flag to indicate if the output file should be
overwritten (if it exists).}

\item{FUN}{Infix function to perform the conversion.}
}
\description{
Convert units from a netCDF file
}
\keyword{internal}
