% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{daily_temp}
\alias{daily_temp}
\title{Calculate daily temperature from \eqn{T_{min}} and \eqn{T_{max}}}
\usage{
daily_temp(
  tmin,
  tmax,
  varid = "tmp",
  timeid = "time",
  latid = "lat",
  lonid = "lon",
  overwrite = TRUE,
  output_filename = NULL
)
}
\arguments{
\item{tmin}{List with \eqn{T_{min}} \code{data} and variable \code{id}.}

\item{tmax}{List with \eqn{T_{max}} \code{data} and variable \code{id}.}

\item{varid}{String with variable ID for daily temperature.}

\item{timeid}{String with the time dimension identifier.}

\item{latid}{String with the latitude dimension identifier.}

\item{lonid}{String with the longitude dimension identifier.}
}
\description{
Calculate daily temperature from \eqn{T_{min}} and \eqn{T_{max}}
}
