% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{days_in_month}
\alias{days_in_month}
\title{Get the days in a month}
\usage{
days_in_month(dates)
}
\arguments{
\item{dates}{Vector of strings with dates.}
}
\value{
Numeric vector with the dates in each month linked to each date.
}
\description{
Get the days in a month from a date string.
}
