% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-date.R
\name{days_month}
\alias{days_month}
\title{Obtain days in each month}
\usage{
days_month(year)
}
\arguments{
\item{year}{Numeric value with a year.}
}
\value{
Named vector with days in each month of \code{year}.
}
\description{
Obtain days in each month
}
\examples{
days_month(2021)
days_month(-11996.92)
}
