% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{E}
\alias{E}
\title{Stomatal sensitivity factor, E (\code{sqrt(Pa)})}
\usage{
E(Tc, beta = 146)
}
\arguments{
\item{Tc}{Numeric value of temperature (°C).}

\item{beta}{Numeric constant, default = 146.}
}
\value{
Numeric value of stomatal sensitivity factor.
}
\description{
Stomatal sensitivity factor, E (\code{sqrt(Pa)})
}
\keyword{internal}
