% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{eta}
\alias{eta}
\title{Viscosity of water}
\usage{
eta(Tc)
}
\arguments{
\item{Tc}{Numeric value of temperature (°C).}
}
\value{
Numeric value of viscosity of water.
}
\description{
Viscosity of water
}
\keyword{internal}
