% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{extract_data}
\alias{extract_data}
\title{Extract data from netCDF file}
\usage{
extract_data(
  filename,
  varid,
  s_year = NULL,
  e_year = NULL,
  timeid = "time",
  latid = "lat",
  lonid = "lon"
)
}
\arguments{
\item{filename}{Filename for the netCDF input (relative or absolute path).}

\item{varid}{String with the main variable identifier.}

\item{s_year}{Numeric value with the start year.}

\item{e_year}{Numeric value with the end year.}

\item{timeid}{String with the time dimension identifier.}

\item{latid}{String with the latitude dimension identifier.}

\item{lonid}{String with the longitude dimension identifier.}
}
\value{
List with dimensions (latitude, longitude, and time) and selected
variable (varid).
}
\description{
Extract data from netCDF file
}
