% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{f}
\alias{f}
\title{f(T, m, ca) (Pa)}
\usage{
f(Tc, MI, co2, scale_factor = 101.325 * 10^-3)
}
\arguments{
\item{Tc}{Numeric value of temperature (°C).}

\item{MI}{Numeric value of moisture index (-).}

\item{co2}{Numeric value of CO2 partial pressure (umol/mol).}

\item{scale_factor}{Scale factor to transform the output, default =
101.325 Pa/ppm at standard sea level pressure.}
}
\value{
Numeric value.
}
\description{
f(T, m, ca) (Pa)
}
\keyword{internal}
