% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-optimisation.R
\name{find_u}
\alias{find_u}
\title{Function used to optimise \eqn{u}}
\usage{
find_u(t0, min_u = -1, max_u = 0.9999999999999, method = "Brent")
}
\arguments{
\item{t0}{Value for \eqn{GDD0/Tmin}, numeric (radians).}

\item{min_u}{Minimum value for \eqn{u (MAT/(MAT-MTCO))}, numeric, unitless.}

\item{max_u}{Maximum value for \eqn{u (MAT/(MAT-MTCO))}, numeric, unitless.}
}
\value{
\eqn{u: MAT / (MAT-MTCO)}.
}
\description{
Function used to optimise \eqn{u}
}
\examples{
codos:::find_u(-4525.758, -1, 0.9999999)
codos:::find_u(-9613.333, -1, 0.9999999)
}
\keyword{internal}
