% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{gdd0}
\alias{gdd0}
\title{Find growing degree days above \code{thr}}
\usage{
gdd0(
  filename,
  varid,
  thr = 0,
  timeid = "time",
  latid = "lat",
  lonid = "lon",
  cpus = 2,
  filter = NULL,
  overwrite = TRUE,
  output_filename = NULL,
  FUN = sum
)
}
\arguments{
\item{filename}{Filename for the netCDF input (relative or absolute path).}

\item{varid}{String with the main variable identifier.}

\item{thr}{Growing season threshold.}

\item{timeid}{String with the time dimension identifier.}

\item{latid}{String with the latitude dimension identifier.}

\item{lonid}{String with the longitude dimension identifier.}

\item{cpus}{Number of CPUs to use for the computation.}

\item{filter}{Variable to be use as filter for the growing season, generally
a structure with temperature data. It must have the same dimensions of
the main variable.}

\item{overwrite}{Boolean flag to indicate if the output file should be
overwritten (if it exists).}

\item{output_filename}{Output filename.}

\item{FUN}{function to apply to get the growing season. Typically
\code{mean} or \code{sum}.}
}
\description{
ind growing degree days above \code{thr} and save output to a netCDF file.
}
