% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-date.R
\name{get_date}
\alias{get_date}
\title{Get a date from a decimal year}
\usage{
get_date(year)
}
\arguments{
\item{year}{Numeric value with a year, include decimals for month and day.}
}
\value{
String with date in the format: \code{Year/Month/Day}
}
\description{
Get a date from a decimal year
}
\examples{
get_date(-11996.92)
get_date(-11996.83)
get_date(-11996.75)
get_date(2021)
get_date(2021.625)
}
