% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{gsl}
\alias{gsl}
\title{Calculate the Growing Season Length (GSL)}
\usage{
gsl(gdd0, mtco)
}
\arguments{
\item{gdd0}{Numeric vector with Growing Degree Days above 0 °C (GDD0) data.}

\item{mtco}{Numeric vector with Mean Temperature of the Coldest Month (MTCO)
data.}
}
\value{
Numeric vector with Growing Season Length (GSL) data.
}
\description{
Calculate the Growing Season Length (GSL) from Growing Degree Days above 0 °C
(GDD0) and Mean Temperature of the Coldest Month (MTCO).
}
