% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interpolation.R
\name{int_acm}
\alias{int_acm}
\title{Mean-preserving autoregressive interpolation}
\usage{
int_acm(y_points, month_len, max_val = ..., min_val = ...)
}
\arguments{
\item{min_val}{}
}
\value{

}
\description{
Perform a mean-preserving autoregressive interpolation.
}
\details{
The idea of how this was done was taken from:
Rymes, M.D. and Myers, D.R., 2001. Mean preserving algorithm for smoothly
interpolating averaged data. Solar Energy, 71(4), pp.225-231.
DOI/URL: https://doi.org/10.1016/S0038-092X(01)00052-4

The method outlined in the paper does not work entirely, and some equations
have been tweaked.
\eqn{(MN(i) = MN(i) + C(K)} as to \eqn{MN(i) - C(K)} and Equation 8 of
the paper.
}
\examples{

}
\references{
Rymes, M.D. and Myers, D.R., 2001. Mean preserving algorithm for smoothly
interpolating averaged data. Solar Energy, 71(4), pp.225-231.
DOI/URL: https://doi.org/10.1016/S0038-092X(01)00052-4
}
\author{
Kamolphat Atsawawaranunt
}
