% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interpolation.R
\name{int_acm2}
\alias{int_acm2}
\title{Simplified version of \code{\link{int_acm}}}
\usage{
int_acm2(y_points, month_len)
}
\arguments{
\item{y_points}{Numeric vector with mean values at each timestep.}

\item{month_len}{Numeric vector with the number of days that each timestep
represents. These can be obtained with \code{\link{days_in_month}} and
\code{\link{retime}}.}
}
\value{
@return Numeric vector with the interpolated values, this one has the same
length as the total sum of month_len.
}
\description{
Simplified version of \code{\link{int_acm}}
}
\keyword{internal}
