% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interpolation.R
\name{int_sin}
\alias{int_sin}
\title{Sinusoidal interpolation}
\usage{
int_sin(minv, maxv, period = 365, x = period, phi = -pi/2)
}
\arguments{
\item{minv}{Numeric value, used as the lower bound.}

\item{maxv}{Numeric value, used as the upper bound.}

\item{period}{Numeric value, period width (e.g. 365 days).}

\item{x}{Numeric value, number of partitions to use.}

\item{phi}{Numeric value, phase shift.}
}
\value{
Numeric vector with the interpolated function. Same length as
\code{x}. Returned invisibly, so it must be assigned to a variable.
}
\description{
Create sinusoidal interpolation based on two values, \code{minv} and
\code{maxv}. The lower and upper bounds/peaks of the function.
}
\examples{
int_sin(-1, 1)
int_sin(-1, 1, period = 10)
}
