% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{julian_day}
\alias{julian_day}
\title{Calculate Julian day}
\usage{
julian_day(year, month, day)
}
\arguments{
\item{year}{Numeric value with the year.}

\item{month}{Numeric value with the month (1-12).}

\item{day}{Numeric value with the day (1-31).}
}
\value{
Numeric value with the Julian day.
}
\description{
Calculate Julian day
}
\examples{
# 13 Aug 2014 (expected 2456882)
julian_day(2014, 8, 13)

}
\references{
Meeus, J. (1991). Astronomical algorithms. 1st ed.
Virginia: Willmann-Bell, Inc. (cit. on pp. 13, 43).
}
