% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{K}
\alias{K}
\title{Effective Michaelis constant of Rubisco (Pa)}
\usage{
K(Tc, dHc = 79430, dHo = 36380, O = 21278, R = 8.314)
}
\arguments{
\item{Tc}{Numeric value of temperature (°C).}

\item{dHc}{Carbon activation energy (J mol^-1), default = 79430.}

\item{dHo}{Oxygen activation energy (J mol^-1), default = 36380.}

\item{O}{Atmospheric concentration of oxygen (Pa), default = 21278.}

\item{R}{Universal gas constant (J mol^-1 K^-1), default = 8.314.}
}
\value{
Numeric value of effective Michaelis constant of Rubisco.
}
\description{
Effective Michaelis constant of Rubisco (Pa)
}
\keyword{internal}
