% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land_mask}
\alias{land_mask}
\title{Land-sea mask dataset}
\format{
A 2D matrix of 720 x 360 (lon x lat).
}
\usage{
data(land_mask)
}
\description{
A dataset containing information of land-sea mask with a grid of 0.5 by 0.5
degrees resolution. \code{TRUE} values represent a land grid cell.
}
\keyword{datasets}
