% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lat}
\alias{lat}
\title{Latitude dataset}
\format{
A list with 4 elements:
\describe{
\item{data}{Numeric vector with the latitude values.}
\item{id}{Identifier to use in netCDF files.}
\item{longname}{String with long name of the variable.}
\item{units}{Latitude units.}
}
}
\usage{
data(lat)
}
\description{
A dataset containing information of the latitude dimension at a resolution
of 0.5 deg.
}
\keyword{datasets}
\keyword{internal}
