% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{lat_lab}
\alias{lat_lab}
\title{Format latitude labels}
\usage{
lat_lab(x)
}
\arguments{
\item{x}{Latitude numeric value.}
}
\value{
String with latitude and corresponding direction suffix.
}
\description{
Format latitude labels
}
\examples{
lat_lab(-30)
lat_lab(30)
}
