% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lon}
\alias{lon}
\title{Longitude dataset}
\format{
A list with 4 elements:
\describe{
\item{data}{Numeric vector with the longitude values.}
\item{id}{Identifier to use in netCDF files.}
\item{longname}{String with long name of the variable.}
\item{units}{Longitude units.}
}
}
\usage{
data(lon)
}
\description{
A dataset containing information of the longitude dimension at a resolution
of 0.5 deg.
}
\keyword{datasets}
\keyword{internal}
