% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{lon_lab}
\alias{lon_lab}
\title{Format longitude labels}
\usage{
lon_lab(x)
}
\arguments{
\item{x}{Longitude numeric value.}
}
\value{
String with longitude and corresponding direction suffix.
}
\description{
Format longitude labels
}
\examples{
lon_lab(-30)
lon_lab(30)
}
