% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interpolation.R
\name{m2d}
\alias{m2d}
\title{Monthly to daily interpolation}
\usage{
m2d(data, time, cpus = 2, thr = 12)
}
\arguments{
\item{data}{Numeric vector with monthly entries.}

\item{time}{Numeric vector with the days in month for a particular time step.}

\item{cpus}{Numeric value with the number of CPUs to use for the computation.}

\item{thr}{Numeric value with the temporal length of the output, in months.}
}
\value{
Numeric value for the daily values interpolated.
}
\description{
Monthly to daily interpolation
}
\keyword{internal}
