% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{mat}
\alias{mat}
\title{Calculate the Mean Annual Temperature (MAT)}
\usage{
mat(gdd0, mtco)
}
\arguments{
\item{gdd0}{Numeric vector with Growing Degree Days above 0 °C (GDD0) data.}

\item{mtco}{Numeric vector with Mean Temperature of the Coldest Month (MTCO)
data.}
}
\value{
Numeric vector with Mean Annual Temperature (MAT) data.
}
\description{
Calculate the Mean Annual Temperature (MAT) from Growing Degree Days above
0 °C (GDD0) and Mean Temperature of the Coldest Month (MTCO).
}
