% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{mi_correction}
\alias{mi_correction}
\title{Calculate corrected moisture index (MI)}
\usage{
mi_correction(T0, T1, MI)
}
\arguments{
\item{T0}{Numeric vector with modern temperature values.}

\item{T1}{Numeric vector with past temperature values.}

\item{MI}{Numeric vector with reconstructed moisture index values.}
}
\value{
Numeric vector with corrected moisture index values.
}
\description{
Calculate corrected moisture index (MI) based on reconstructed MI and
temperature, past and modern.
}
\examples{
codos::mi_correction(11.5795742, 12.36931467, 0.330794535)
}
