% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{monthly_clim}
\alias{monthly_clim}
\title{Create monthly climatology}
\usage{
monthly_clim(
  filename,
  varid,
  s_year,
  e_year,
  timeid = "time",
  latid = "lat",
  lonid = "lon",
  overwrite = TRUE
)
}
\arguments{
\item{filename}{Filename for the netCDF input (relative or absolute path).}

\item{varid}{String with the main variable identifier.}

\item{s_year}{Numeric value with the start year.}

\item{e_year}{Numeric value with the end year.}

\item{timeid}{String with the time dimension identifier.}

\item{latid}{String with the latitude dimension identifier.}

\item{lonid}{String with the longitude dimension identifier.}

\item{overwrite}{Boolean flag to indicate if the output file should be
overwritten (if it exists).}
}
\description{
Create monthly climatology
}
