% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{nc2ts}
\alias{nc2ts}
\title{Convert netCDF to time series}
\usage{
nc2ts(
  filename,
  varid,
  timeid = "time",
  latid = "lat",
  lonid = "lon",
  plot = TRUE
)
}
\arguments{
\item{filename}{Filename for the netCDF input (relative or absolute path).}

\item{varid}{String with the main variable identifier.}

\item{timeid}{String with the time dimension identifier.}

\item{latid}{String with the latitude dimension identifier.}

\item{lonid}{String with the longitude dimension identifier.}

\item{plot}{Boolean flag to indicate whether a plot for the time series
should be generated.}
}
\value{
Tibble with the time and mean values.
}
\description{
Convert netCDF file to a time series using the area-weighted mean (based on
the latitudes in the netCDF file).
}
\keyword{internal}
