% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{nc_check}
\alias{nc_check}
\title{Check netCDF file}
\usage{
nc_check(filename, varid, timeid, latid, lonid)
}
\arguments{
\item{filename}{Filename for the netCDF input (relative or absolute path).}

\item{varid}{String with the main variable identifier.}

\item{timeid}{String with the time dimension identifier.}

\item{latid}{String with the latitude dimension identifier.}

\item{lonid}{String with the longitude dimension identifier.}
}
\value{
Reference to \code{ncdf4} object.
}
\description{
Check netCDF file
}
\keyword{internal}
