% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{nc_gs}
\alias{nc_gs}
\title{Find growing season and save netCDF file}
\usage{
nc_gs(
  filename,
  varid,
  thr = 0,
  timeid = "time",
  latid = "lat",
  lonid = "lon",
  cpus = 2,
  overwrite = TRUE
)
}
\arguments{
\item{filename}{Filename for the netCDF input (relative or absolute path).}

\item{varid}{String with the main variable identifier.}

\item{thr}{Growing season threshold.}

\item{timeid}{String with the time dimension identifier.}

\item{latid}{String with the latitude dimension identifier.}

\item{lonid}{String with the longitude dimension identifier.}

\item{cpus}{Number of CPUs to use for the computation.}

\item{overwrite}{Boolean flag to indicate if the output file should be
overwritten (if it exists).}
}
\description{
Find growing season and save netCDF file
}
