% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{nc_int}
\alias{nc_int}
\title{Interpolate netCDF file}
\usage{
nc_int(
  filename,
  varid,
  timeid = "time",
  latid = "lat",
  lonid = "lon",
  cpus = 2,
  s_year = 1961,
  overwrite = TRUE
)
}
\arguments{
\item{filename}{Filename for the netCDF input (relative or absolute path).}

\item{varid}{String with the main variable identifier.}

\item{timeid}{String with the time dimension identifier.}

\item{latid}{String with the latitude dimension identifier.}

\item{lonid}{String with the longitude dimension identifier.}

\item{cpus}{}

\item{s_year}{Numeric value with the start year.}

\item{overwrite}{Boolean flag to indicate if the output file should be
overwritten (if it exists).}
}
\description{
Interpolate netCDF file
}
