% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{nc_mi}
\alias{nc_mi}
\title{Calculate soil moisture index and save netCDF file}
\usage{
nc_mi(filename, pet, pre, lat = NULL, lon = NULL, cpus = 2, overwrite = TRUE)
}
\arguments{
\item{filename}{String with the output filename (.nc).}

\item{pet}{3D structure with potential evapotranspiration data. These values
can be calculated with the function \code{\link{splash_evap}}.}

\item{lat}{List with latitude \code{data} and variable \code{id}.}

\item{lon}{List with longitude \code{data} and variable \code{id}.}

\item{cpus}{Number of CPUs to use for the computation.}

\item{overwrite}{Boolean flag to indicate if the output file should be
overwritten (if it exists).}

\item{tmn}{3D structure with minimum temperature data.}

\item{tmx}{3D structure with maximum temperature data.}
}
\description{
Calculate soil moisture index and save netCDF file
}
