% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_save.R
\name{nc_save_timeless}
\alias{nc_save_timeless}
\title{Save data as netCDF file without a time dimension}
\usage{
nc_save_timeless(filename, var, lat, lon, var_atts = NULL, overwrite = TRUE)
}
\arguments{
\item{filename}{String with output filename with or without path.}

\item{var}{List with data and metadata for the main variable.}

\item{lat}{List with data and metadata for the latitude dimension.}

\item{lon}{List with data and metadata for the longitude dimension.}

\item{var_atts}{List with extra attributes for the main variable.}

\item{overwrite}{Boolean flag to indicate if the file should be overwritten.
(if exists).}
}
\description{
Save data as netCDF file without a time dimension
}
\keyword{internal}
