% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{nc_Tg}
\alias{nc_Tg}
\title{Wrapper for \code{\link{T_g}}}
\usage{
nc_Tg(
  filename,
  dcl,
  tmn,
  tmx,
  lat = NULL,
  lon = NULL,
  cpus = 2,
  overwrite = TRUE
)
}
\arguments{
\item{filename}{String with the output filename (.nc).}

\item{dcl}{3D structure with solar declination angle data.}

\item{tmn}{3D structure with minimum temperature data.}

\item{tmx}{3D structure with maximum temperature data.}

\item{lat}{List with latitude \code{data} and variable \code{id}.}

\item{lon}{List with longitude \code{data} and variable \code{id}.}

\item{cpus}{Number of CPUs to use for the computation.}

\item{overwrite}{Boolean flag to indicate if the output file should be
overwritten (if it exists).}
}
\description{
Wrapper for \code{\link{T_g}} (mean daytime air temperature).
}
