% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{past_co2_loess}
\alias{past_co2_loess}
\title{Find past CO2 using \code{loess}}
\usage{
past_co2_loess(age, ref = codos::ice_core, span = 0.1, ...)
}
\arguments{
\item{age}{Numeric value with the \code{age}.}

\item{ref}{Reference data frame containing ice core composite
information, defaults to \code{codos::ice_core}:

Bereiter, B., Eggleston, S., Schmitt, J., Nehrbass‐Ahles, C., Stocker, T. F.,
Fischer, H., Kipfstuhl, S., and Chappellaz, J. (2015), Revision of the EPICA
Dome C CO2 record from 800 to 600 kyr before present, Geophys. Res. Lett.,
42, 542– 549, \url{doi:10.1002/2014GL061957}.}

\item{span}{Numeric value, \eqn{\alpha}, which controls the degree of
smoothing.}

\item{...}{Extra arguments passed to \code{\link[stats:loess]{stats::loess}}.}
}
\value{
Numeric value of CO2 linked to the given \code{age}.
}
\description{
Find past CO2 linked to a given \code{age} using \code{loess}.
}
\keyword{internal}
