% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{plot_map}
\alias{plot_map}
\title{Create simple map}
\usage{
plot_map(data, lat, lon)
}
\arguments{
\item{data}{2D array with the data to be mapped.}

\item{lat}{Numeric array with latitude data (y-axis).}

\item{lon}{Numeric array with longitude data (x-axis).}
}
\value{
Graphical object.
}
\description{
Create simple map
}
\keyword{internal}
