% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{R_n}
\alias{R_n}
\title{Annual net radiation at the vegetated surface}
\usage{
R_n(self, Temp, m, lat = -30)
}
\value{
Annual net radiation at the vegetated surface.
}
\description{
Annual net radiation at the vegetated surface
}
\details{
Given by (2.4):
\eqn{R_n = 0.83 R_0 (0.25 + 0.5 S_f) - (107 - T)(0.2 + 0.8 S_f)}.
}
\references{
I.C. Prentice, S.F. Cleator, Y.H. Huang, S.P. Harrison, I. Roulstone,
"Reconstructing ice-age palaeoclimates: Quantifying low-CO2 effects on
plants", Global and Planetary Change, Volume 149, 2017, Pages 166-176,
DOI: \url{https://doi.org/10.1016/j.gloplacha.2016.12.012}.
}
