% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{retime}
\alias{retime}
\title{Change time axis}
\usage{
retime(time_var, ref_date = lubridate::date("1870-01-01"), duration = "months")
}
\arguments{
\item{time_var}{Numeric array with current time axis values.}

\item{ref_date}{Reference data for the current axis.}

\item{duration}{Interval between entries.}
}
\value{

}
\description{
Change time axis from a reference date.
For example "Months since 1870-01-01", set \code{ref_date = "1870-01-01"}
and \code{duration = "months"}.
}
