% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{S_f}
\alias{S_f}
\title{Fraction of sunshine hours}
\usage{
S_f(m)
}
\arguments{
\item{m}{Numeric value for moisture index.}
}
\value{
Fraction of sunshine hours.
}
\description{
Fraction of sunshine hours
}
\details{
Given by (2.5):
\eqn{S_f = 0.6611 e^{-0.74m} + 0.2175}.
}
\references{
I.C. Prentice, S.F. Cleator, Y.H. Huang, S.P. Harrison, I. Roulstone,
"Reconstructing ice-age palaeoclimates: Quantifying low-CO2 effects on
plants", Global and Planetary Change, Volume 149, 2017, Pages 166-176,
DOI: \url{https://doi.org/10.1016/j.gloplacha.2016.12.012}.
}
\keyword{internal}
