% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interpolation.R
\name{shift}
\alias{shift}
\title{Shift vector}
\usage{
shift(x, n, invert = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{n}{Number of positions.}

\item{invert}{Boolean flag to indicate if the vector should be inverted.}
}
\value{
Shifted numeric vector.
}
\description{
Shift vector by \code{n} positions.
}
\examples{
shift(c(1:10), 2)

}
\references{
https://stackoverflow.com/questions/26997586/shifting-a-vector
}
\author{
Kamolphat Atsawawaranunt
}
