% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-splash.R
\name{splash_dcl}
\alias{splash_dcl}
\title{Calculate solar declination
Calculate solar declination angle in degrees using SPLASH V1.0:
\url{https://doi.org/10.5281/zenodo.376293}.}
\usage{
splash_dcl(year)
}
\arguments{
\item{year}{Numeric value with the year.}
}
\value{
Numeric vector with the solar declination angle for each day of the
given year.
}
\description{
Calculate solar declination
Calculate solar declination angle in degrees using SPLASH V1.0:
\url{https://doi.org/10.5281/zenodo.376293}.
}
\examples{
splash_dcl(1960)
splash_dcl(1961)
}
