% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-splash.R
\name{splash_solar}
\alias{splash_solar}
\title{Calculate solar declination
Calculate solar declination angle in degrees using SPLASH:
\url{https://doi.org/10.5281/zenodo.376293}.}
\usage{
splash_solar(
  filename,
  elv,
  sf,
  tmp,
  year,
  lat = NULL,
  lon = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{filename}{String with the output filename (.nc).}

\item{elv}{2D structure with elevation data.}

\item{sf}{3D structure with sunshine fraction data.}

\item{tmp}{3D structure with daily temperature data.}

\item{year}{Numeric value with the year.}

\item{lat}{List with latitude \code{data} and variable \code{id}.}

\item{lon}{List with longitude \code{data} and variable \code{id}.}

\item{overwrite}{Boolean flag to indicate if the output file should be
overwritten (if it exists).}
}
\description{
Calculate solar declination
Calculate solar declination angle in degrees using SPLASH:
\url{https://doi.org/10.5281/zenodo.376293}.
}
