% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{T_g}
\alias{T_g}
\title{Mean daytime air temperature}
\usage{
T_g(lat, delta, Temp_max, Temp_min)
}
\arguments{
\item{lat}{Latitude (\eqn{\lambda}).}

\item{delta}{Monthly average solar declination (\eqn{\delta}).}

\item{Temp_max}{Maximum value of temperature (\eqn{T_{max}}).}

\item{Temp_min}{Minimum value of temperature (\eqn{T_{min}}).}
}
\value{
Mean daytime air temperature.
}
\description{
Mean daytime air temperature (\eqn{T_g}) was estimated for each month by
assuming the diurnal temperature cycle to follow a sine curve, with daylight
hours determined by latitude and month.
}
\details{
Given by (7):

\eqn{T_g = T_{max}\left[\frac{1}{2} +
                        \frac{(1-x^2)^{1/2}}{2} \cos^{-1}{x}\right] +
           T_{min}\left[\frac{1}{2} + \frac{(1-x^2)^{1/2}}{2} \cos^{-1}{x}\right]}

where

\eqn{x = -\tan{\lambda}\tan{\delta}}.
}
