% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{ts_comp}
\alias{ts_comp}
\title{Compare time series}
\usage{
ts_comp(
  climatology,
  interpolated,
  month_len,
  vars = c("cld", "pre", "tmn", "tmx", "vap"),
  main = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{climatology}{Numeric vector with monthly climatologies data.}

\item{interpolated}{Numeric vector with daily interpolated data.}

\item{month_len}{Numeric vector with the number of days in each month.}

\item{vars}{Vector of strings with variables to be plotted.}

\item{main}{String with title for the plot.}

\item{xlab}{String with label for the x-axis.}

\item{ylab}{String with label for the y-axis.}
}
\value{
\code{ggplot2} graphic object.
}
\description{
Compare time series for climatologies (monthly) and interpolated (daily)
data.
}
