% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{ts_plot}
\alias{ts_plot}
\title{Plot time series}
\usage{
ts_plot(
  data,
  vars = c("cld", "pre", "tmn", "tmx", "vap"),
  days = 365,
  months = 12,
  main = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{data}{Numeric vector with the data.}

\item{vars}{Vector of strings with variables to be plotted.}

\item{days}{Numeric, days in a year.}

\item{months}{Numeric, months in a year.}

\item{main}{String with title for the plot.}

\item{xlab}{String with label for the x-axis.}

\item{ylab}{String with label for the y-axis.}
}
\value{
\code{ggplot2} graphic object.
}
\description{
Plot time series for multiple variables.
}
