% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{vpd_internal}
\alias{vpd_internal}
\title{Vapour-pressure deficit (Pa)}
\usage{
vpd_internal(Tc, MI, scale_factor = 100)
}
\arguments{
\item{Tc}{Numeric value of temperature (°C).}

\item{MI}{Numeric value of moisture index.}

\item{scale_factor}{Scale factor to transform the output, default =
100 Pa/hPa.}
}
\value{
Numeric value of vapour-pressure deficit.
}
\description{
Vapour-pressure deficit (Pa)
}
\keyword{internal}
