% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\docType{package}
\name{codos-package}
\alias{codos}
\alias{codos-package}
\title{codos: CO2 Correction Tools}
\description{
CO2 Correction Tools.
}
\author{
\strong{Maintainer}: Roberto Villegas-Diaz \email{r.villegas-diaz@reading.ac.uk} (\href{https://orcid.org/0000-0001-5036-8661}{ORCID})

Authors:
\itemize{
  \item Sandy P. Harrison \email{s.p.harrison@reading.ac.uk} (\href{https://orcid.org/0000-0001-5687-1903}{ORCID})
  \item Iain Colin Prentice \email{c.prentice@imperial.ac.uk} (\href{https://orcid.org/0000-0002-1296-6764}{ORCID})
}

Other contributors:
\itemize{
  \item SPECIAL Research Group @ University of Reading [copyright holder]
}

}
\keyword{internal}
