% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{corrected_mi}
\alias{corrected_mi}
\title{Calculate corrected moisture index (MI)}
\usage{
corrected_mi(Tc0, Tc1, MI, ca0, ca1, ...)
}
\arguments{
\item{Tc0}{Numeric vector with modern mean growing season temperature
values (°C).}

\item{Tc1}{Numeric vector with palaeo (past) mean growing season temperature
values (°C).}

\item{MI}{Numeric vector with reconstructed moisture index values (-).}

\item{ca0}{Numeric vector with modern ('recent') CO2 partial pressures
(umol/mol).}

\item{ca1}{Numeric vector with palaeo (past) CO2 partial pressures
(umol/mol).}

\item{...}{
  Arguments passed on to \code{\link[=vpd]{vpd}}
  \describe{
    \item{\code{scale_factor}}{Scale factor to transform the output, default =
101.325 Pa/ppm at standard sea level pressure.}
  }}
}
\value{
Numeric vector with corrected moisture index values.
}
\description{
Calculate corrected moisture index (MI) based on reconstructed MI, CO2 and
temperature, past and modern.
}
