% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{eta}
\alias{eta}
\title{Viscosity of water}
\usage{
eta(self, Temp)
}
\arguments{
\item{self}{Reference to class \code{p_model}.}

\item{Temp}{Temperature.}
}
\value{
Annual net radiation at the vegetated surface.

Viscosity of water.
}
\description{
Viscosity of water
}
\details{
Given by (2.8):
\eqn{\eta = 0.024258 exp{\frac{580}{T_K} - 183}}.
}
\references{
I.C. Prentice, S.F. Cleator, Y.H. Huang, S.P. Harrison, I. Roulstone,
"Reconstructing ice-age palaeoclimates: Quantifying low-CO2 effects on
plants", Global and Planetary Change, Volume 149, 2017, Pages 166-176,
DOI: \url{https://doi.org/10.1016/j.gloplacha.2016.12.012}.
}
