% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R, R/utils-optimisation.R
\name{f}
\alias{f}
\title{f(T, m, ca) (Pa)}
\usage{
f(u, t0)

f(u, t0)
}
\arguments{
\item{u}{\eqn{u}.}

\item{t0}{\eqn{GDD0/Tmin}.}

\item{Tc}{Numeric value of temperature (°C).}

\item{MI}{Numeric value of moisture index (-).}

\item{co2}{Numeric value of CO2 partial pressure (umol/mol).}

\item{scale_factor}{Scale factor to transform the output, default =
101.325 Pa/ppm at standard sea level pressure.}
}
\value{
Numeric value.

Numeric value
}
\description{
f(T, m, ca) (Pa)

Function to calculate \eqn{u}
}
\details{
This function is used to optimise values x and y, for the equation to
calculate MAT based on GDD0 and MTCO.

This equation is only used when MTCO < 0 and GDD0 > 0

If MTCO > 0, GDD0 must be > 0, T0 = GDD0 / 2 * pi
Units:
GDD0 = K rad

u = MAT/(MAT-MTCO)
}
\keyword{internal}
