% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-optimisation.R
\name{f}
\alias{f}
\title{Function to calculate \eqn{u}}
\usage{
f(u, t0)
}
\arguments{
\item{u}{\eqn{u}.}

\item{t0}{\eqn{GDD0/Tmin}.}
}
\value{

}
\description{
Function to calculate \eqn{u}
}
\details{
This function is used to optimise for values x and y, based on the equation
for MAT based on GDD0 and MTCO.

This equation is only used when MTCO < 0 and GDD0 > 0

If MTCO > 0, GDD0 must be > 0, T0 = GDD0/2*pi
Units:
GDD0 = K rad

u = MAT/(MAT-MTCO)
}
\keyword{internal}
