% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codos.R
\name{K}
\alias{K}
\title{Effective Michaelis-Menten coefficient for carboxylation}
\usage{
K(self, Temp)
}
\arguments{
\item{self}{Reference to class \code{p_model}.}

\item{Temp}{Temperature.}
}
\value{
Annual net radiation at the vegetated surface.

Viscosity of water.
}
\description{
Effective Michaelis-Menten coefficient for carboxylation
}
\details{
Given by (2.9):

\eqn{K = K_C \left(1 + \frac{O}{K_O}\right)}.

where

\eqn{K_C = 404.9 \exp\left[\frac{\Delta H_C}{R}\left(\frac{1}{298} - \frac{1}{T_K}\right)\right]}

and

\eqn{K_O = 278.4 \exp\left[\left(\frac{\Delta H_O}{R}\right)\left(\frac{1}{298} - \frac{1}{T_K}\right)\right]}.
}
\references{
I.C. Prentice, S.F. Cleator, Y.H. Huang, S.P. Harrison, I. Roulstone,
"Reconstructing ice-age palaeoclimates: Quantifying low-CO2 effects on
plants", Global and Planetary Change, Volume 149, 2017, Pages 166-176,
DOI: \url{https://doi.org/10.1016/j.gloplacha.2016.12.012}.
}
