% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land_mask}
\alias{land_mask}
\title{CRU TS 4.04 Landmask dataset}
\format{
A 2D matrix of 720 x 360 (lon x lat) with boolean values:
\describe{
\item{\code{TRUE}}{Land grid cell.}
\item{\code{FALSE}}{Non-land grid cell.}
}
}
\usage{
data(land_mask)
}
\description{
A dataset containing information of land mask with a grid of 0.5 by 0.5
degrees resolution. \code{TRUE} values represent a land grid cell.
}
\details{
The mask was generated by looking at the values of \code{tmn} and \code{vap},
for each grid cell, if at least one timestep had a non-missing value, then
the entire grid cell was marked as a land grid cell.
}
\references{
University of East Anglia Climatic Research Unit; Harris, I.C.; Jones, P.D.;
Osborn, T. (2020): CRU TS4.04: Climatic Research Unit (CRU) Time-Series (TS)
version 4.04 of high-resolution gridded data of month-by-month variation in
climate (Jan. 1901- Dec. 2019). Centre for Environmental Data Analysis.
\url{https://catalogue.ceda.ac.uk/uuid/89e1e34ec3554dc98594a5732622bce9}
}
\keyword{datasets}
\keyword{internal}
