% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nc.R
\name{nc_var_get}
\alias{nc_var_get}
\title{Get variable from netCDF file}
\usage{
nc_var_get(filename, varid, is.dim = FALSE)
}
\arguments{
\item{filename}{Filename for the netCDF input (relative or absolute path).}

\item{varid}{String with the main variable identifier.}

\item{is.dim}{Boolean flag to indicate if the variable is a dimension
(e.g. time, latitude, or longitude).}
}
\value{
List with data, filename, id, and units linked to the variable.
}
\description{
Get variable from netCDF file
}
\keyword{internal}
